==> 05_Get_30Y_Mortgage_Rate.r <==
##-----------------------------------------------------------------------
##
##-----------------------------------------------------------------------

##-----------------------------------------------------------------------
## clean the cache
##-----------------------------------------------------------------------
#rm(list=ls())

##-----------------------------------------------------------------------
## load libraries
##-----------------------------------------------------------------------
source("/Users/alexstephens/Development/fnma/fnma_code/99_Load_Libraries.r")

##-----------------------------------------------------------------------
## set the working directory
##-----------------------------------------------------------------------
setwd("/Users/alexstephens/Development/fnma/data/rate")
rateDirectory <- "/Users/alexstephens/Development/fnma/data/rate"
rate.out      <- "FRED_30Y_Conventional_Mortgage_Rate.Rda"

##-----------------------------------------------------------------------
## read the raw data
##-----------------------------------------------------------------------
rate.data     <- as.data.table(read.csv(file="FRED_30Y_Conventional_Mortgage_Rate.csv", sep=",", header=TRUE))

##-----------------------------------------------------------------------
## clean series_id and define the key
##-----------------------------------------------------------------------
rate.data[,date := as.Date(observation_date,"%Y-%m-%d")]
rate.data[,c("yy","mm") := list(year(date),ifelse(month(date)<10,paste0("0",month(date)),month(date)))]

##-----------------------------------------------------------------------
## create the final seried; define a key
##-----------------------------------------------------------------------
rate = rate.data

##-----------------------------------------------------------------------
## create lagged variables
##-----------------------------------------------------------------------
setorderv(rate, c("date"))

##-----------------------------------------------------------------------
## define keys
##-----------------------------------------------------------------------
rate[ , Key := paste0(yy,mm)]
setkey(rate, Key)

##-----------------------------------------------------------------------
## save
##-----------------------------------------------------------------------
save(rate, file=paste0(rateDirectory,"/",rate.out))
