==> 99_FNMA_Raw_Data_Attributes.r <==
##-----------------------------------------------------------------------
## Define Acquisition File Variables & Column Classes
##-----------------------------------------------------------------------
Acquisitions_Variables <-
c(  "LOAN_ID",
    "ORIG_CHN",
    "Seller.Name",
    "ORIG_RT",
    "ORIG_AMT",
    "ORIG_TRM",
    "ORIG_DTE",
    "FRST_DTE",
    "OLTV",
    "OCLTV",
    "NUM_BO",
    "DTI",
    "CSCORE_B",
    "FTHB_FLG",
    "PURPOSE",
    "PROP_TYP",
    "NUM_UNIT",
    "OCC_STAT",
    "STATE",
    "ZIP_3",
    "MI_PCT",
    "Product.Type",
    "CSCORE_C")

Acquisition_ColClasses <-
c(  "character",
    "character",
    "character",
    "numeric",
    "numeric",
    "integer",
    "character",
    "character",
    "numeric",
    "numeric",
    "character",
    "numeric",
    "numeric",
    "character",
    "character",
    "character",
    "character",
    "character",
    "character",
    "character",
    "numeric",
    "character",
    "numeric")

##-----------------------------------------------------------------------
## Define Performance File Variables & Column Classes
##-----------------------------------------------------------------------
Performance_Variables <-
    c(  "LOAN_ID",
        "Monthly.Rpt.Prd",
        "Servicer.Name",
        "LAST_RT",
        "LAST_UPB",
        "Loan.Age",
        "Months.To.Legal.Mat",
        "Adj.Month.To.Mat",
        "Maturity.Date",
        "MSA",
        "Delq.Status",
        "MOD_FLAG",
        "Zero.Bal.Code",
        "ZB_DTE",
        "LPI_DTE",
        "FCC_DTE",
        "DISP_DT",
        "FCC_COST",
        "PP_COST",
        "AR_COST",
        "IE_COST",
        "TAX_COST",
        "NS_PROCS",
        "CE_PROCS",
        "RMW_PROCS",
        "O_PROCS",
        "NON_INT_UPB",
        "PRIN_FORG_UPB")

Performance_ColClasses <-
    c(  "character",
        "character",
        "character",
        "numeric",
        "numeric",
        "numeric",
        "numeric",
        "numeric",
        "character",
        "character",
        "character",
        "character",
        "character",
        "character",
        "character",
        "character",
        "character",
        "numeric",
        "numeric",
        "numeric",
        "numeric",
        "numeric",
        "numeric",
        "numeric",
        "numeric",
        "numeric",
        "numeric",
        "numeric")
